Add service component
The @Injectable() decorator tells TypeScript to emit metadata about the service. The metadata specifies that Angular may need to inject other dependencies into this service.

Don't use new with the HeroService

To teach the injector how to make a HeroService, add the following providers array property to the bottom of the component metadata in the @Component call.

The ngOnInit lifecycle hook
Write an ngOnInit method with the initialization logic inside. Angular will call it at the right time. In this case, initialize by calling getHeroes().

Async services and Promises
Pass the callback function as an argument to the Promise's then() method: