Splitting the AppComponent
The current app loads AppComponent and immediately displays the list of heroes.

The revised app should present a shell with a choice of views (Dashboard and Heroes) and then default to one of them.

The AppComponent should only handle navigation, so you'll move the display of Heroes out of AppComponent and into its own HeroesComponent.

Routing is another name for navigation. The router is the mechanism for navigating from view to view.